Fix pull to only run when changes are found

+++ ellipsis.sh
@@ -10,14 +10,18 @@ pkg.install() {
 }
 
 pkg.pull() {
-    # Unlink old files
-    hooks.unlink
+    # Check for updates on git
+    git remote update 2>&1 > /dev/null
+    if git.is_behind; then
+        # Unlink old files
+        hooks.unlink
 
-    # Pull down package changes
-    git.pull
+        # Pull down package changes
+        git.pull
 
-    # Link files
-    pkg.link
+        # Link files
+        pkg.link
+    fi
 
     # Configure git
     bash $PKG_PATH/run.sh "$ELLIPSIS_SRC"
